+++ doc/CHANGES.md
@@ -1,6 +1,7 @@
 2016-09-02  Release 1.58.0 (to be released)
 
   * Roll back overriding the url of a metric (release 1.57.0).
+  * The OWASP dependency warnings metrics now link to the HTML report when using the XML report as source.
   
   
 2016-09-01  Release 1.57.3